AMBARI-1447. Report current Stack version for all host components. (Sumit Mohanty via mahadev)
git-svn-id: https://svn.apache.org/repos/asf/incubator/ambari/trunk@1449806 13f79535-47bb-0310-9956-ffa450edef68

+++ ambari-server/src/main/java/org/apache/ambari/server/orm/dao/ExecutionCommandDAO.java
@@ -51,7 +51,7 @@ public class ExecutionCommandDAO {
   }
 
   @Transactional
-  public void removeByPK(int taskId) {
+  public void removeByPK(long taskId) {
     remove(findByPK(taskId));
   }
 }
